+++ index.html
@@ -5,7 +5,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src='js/main.js'></script>
+  <script src="js/main.js" type="module="></script>
   <title>Кексобукинг</title>
 </head>
 <body>
